agents/requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
openai==2.15.0
langchain==1.2.3
langchain-openai==1.1.7
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.7.3